Added Forms.Lua and Inputs.Lua This will allow more spaced out code
