+++ Lua/NES/MarioAI.lua
@@ -145,187 +145,7 @@ NoveltyConstant=1
 
 CurrentNSFitness=0
 
-
---[[
-GetPostions: Return the postion of Mario Using in game Hex Bits
---]]
-function getPositions()
-
-	--If statements for which mario game
-	if gameinfo.getromname() == "Super Mario World (USA)" then
-
-		--Mario X and Y bits are Both 2 Byte Words
-		marioX = memory.read_s16_le(0x94)
-		marioY = memory.read_s16_le(0x96)
-		local layer1x = memory.read_s16_le(0x1A);
-		local layer1y = memory.read_s16_le(0x1C);
-
-		screenX = marioX-layer1x
-		screenY = marioY-layer1y
-	elseif gameinfo.getromname() == "Super Mario Bros." then
-
-		--In the classic game the X bit is done using a paging bit first
-		--This allows you to have bigger numbers than 256
-		marioX = memory.readbyte(0x6D) * 0x100 + memory.readbyte(0x86)
-		marioY = memory.readbyte(0x03B8)+16
-		marioScore = memory.readbyte(0x7D8)*100000+memory.readbyte(0x07D9)*10000+memory.readbyte(0x07DA)*1000
-		marioScore = marioScore+memory.readbyte(0x07DB)*100 + memory.readbyte(0x07DC)*10 + memory.readbyte(0x07DC)*1
-		marioLevel = memory.readbyte(0x075C)
-		marioWorld = memory.readbyte(0x075F)
-		--marioScore= 19
-		screenX = memory.readbyte(0x03AD)
-		screenY = memory.readbyte(0x03B8)
-	end
-end
-
-
---[[
-getTile: Return each of the tiles used for the input
-Return:  Whether the tile is
---]]
-function getTile(dx, dy)
-	if gameinfo.getromname() == "Super Mario World (USA)" then
-
-		--Find the exact tile and get the offset
-		x = math.floor((marioX+dx+8)/16)
-		y = math.floor((marioY+dy)/16)
-
-		--Read the byte of that dx,dy
-		return memory.readbyte(0x1C800 + math.floor(x/0x10)*0x1B0 + y*0x10 + x%0x10)
-	elseif gameinfo.getromname() == "Super Mario Bros." then
-
-		--Dx plus what page they are on
-		local x = marioX + dx + 8
-		local y = marioY + dy - 16
-		local page = math.floor(x/256)%2
-
-
-		local subx = math.floor((x%256)/16)
-		local suby = math.floor((y - 32)/16)
-		local addr = 0x500 + page*13*16+suby*16+subx
-
-
-		if suby >= 13 or suby < 0 then
-			return 0
-		end
-
-		if memory.readbyte(addr) ~= 0 then
-			return 1
-		else
-			return 0
-		end
-	end
-end
-
---[[
-GetSprites: Return the monster if they are on a tile
---]]
-function getSprites()
-	if gameinfo.getromname() == "Super Mario World (USA)" then
-		local sprites = {}
-		for slot=0,11 do
-			local status = memory.readbyte(0x14C8+slot)
-			if status ~= 0 then
-				spritex = memory.readbyte(0xE4+slot) + memory.readbyte(0x14E0+slot)*256
-				spritey = memory.readbyte(0xD8+slot) + memory.readbyte(0x14D4+slot)*256
-				sprites[#sprites+1] = {["x"]=spritex, ["y"]=spritey}
-			end
-		end
-
-		return sprites
-	elseif gameinfo.getromname() == "Super Mario Bros." then
-		local sprites = {}
-		for slot=0,4 do
-			--Reads the four slots and see if there are monsters 000F-0012
-			local enemy = memory.readbyte(0xF+slot)
-			--What page number an enemy is on and what's it's x and y
-			if enemy ~= 0 then
-				local ex = memory.readbyte(0x6E + slot)*0x100 + memory.readbyte(0x87+slot)
-				local ey = memory.readbyte(0xCF + slot)+24
-				--Go to the end of the list no append
-				--Add the tuple X and Y of each monster
-				sprites[#sprites+1] = {["x"]=ex,["y"]=ey}
-			end
-		end
-
-		return sprites
-	end
-end
-
-
---[[
-GetExtendedSprites: Only used for SNES
---]]
-function getExtendedSprites()
-	if gameinfo.getromname() == "Super Mario World (USA)" then
-		local extended = {}
-		for slot=0,11 do
-			local number = memory.readbyte(0x170B+slot)
-			if number ~= 0 then
-				spritex = memory.readbyte(0x171F+slot) + memory.readbyte(0x1733+slot)*256
-				spritey = memory.readbyte(0x1715+slot) + memory.readbyte(0x1729+slot)*256
-				extended[#extended+1] = {["x"]=spritex, ["y"]=spritey}
-			end
-		end
-
-		return extended
-	elseif gameinfo.getromname() == "Super Mario Bros." then
-		return {}
-	end
-end
-
-
---[[
-GetInputs: Uses the previous functions to generates all inputs
---]]
-function getInputs()
-
-
-	getPositions()
-
-	sprites = getSprites()
-	extended = getExtendedSprites()
-
-	local inputs = {}
-	--For loop dy from negative 16 times the BoxRadius to postive 16 times the BoxRadius incrementing by 16
-	--It will loop BoxRadius*Two
-	for dy=-BoxRadius*16,BoxRadius*16,16 do
-		--Do the same thing for y's
-		for dx=-BoxRadius*16,BoxRadius*16,16 do
-			--Initializes all inputs to 0
-			inputs[#inputs+1] = 0
-
-
-			tile = getTile(dx, dy)
-			--If the current location is a block set input to 1
-			if tile == 1 and marioY+dy < 0x1B0 then
-				inputs[#inputs] = 1
-			end
-			--If the current tile is then where a monster is set it to -1
-			for i = 1,#sprites do
-				distx = math.abs(sprites[i]["x"] - (marioX+dx))
-				disty = math.abs(sprites[i]["y"] - (marioY+dy))
-				if distx <= 8 and disty <= 8 then
-					inputs[#inputs] = -1
-				end
-			end
-			--Only for SNES
-			for i = 1,#extended do
-				distx = math.abs(extended[i]["x"] - (marioX+dx))
-				disty = math.abs(extended[i]["y"] - (marioY+dy))
-				if distx < 8 and disty < 8 then
-					inputs[#inputs] = -1
-				end
-			end
-		end
-	end
-
-	--mariovx = memory.read_s8(0x7B)
-	--mariovy = memory.read_s8(0x7D)
-
-	return inputs
-end
-
+require "Inputs"
 
 --[[
 Sigmoid: Determine if a neuron value is postive or negative
@@ -1840,88 +1660,9 @@ function CalculateSpeciesCord(species,genome)
 	return species*100+genome
 end
 
--- function CheckLandscape(cordLocation,cordSpecies)
--- 	if pool.landscape[tostring(cordLocation)]=nil then
--- 		console.writeline("new")
--- 	end
--- end
 
-function incrementX(Amount)
-	xvalue=xvalue+Amount
-	return xvalue
-end
-
-function incrementY(Amount)
-	yvalue=yvalue+Amount
-	return yvalue
-end
-
---Create Fitness Form
-function FitnessBox(AmountX,AmountY,BoxX,BoxY)
-yvalue=5 xvalue=5
-form = forms.newform(BoxX, BoxY, "Fitness")
-
-showNetwork = forms.checkbox(form, "Show Map", xvalue, yvalue, "true") --A checkbox to see whether or not the eye(inputs) and controls(outputs) is shown
-showMutationRates = forms.checkbox(form, "Show Mutate", incrementX(AmountX), yvalue)
-yvalue=incrementY(AmountY) xvalue=0
---Label to describe the differnt types of fitness
-FitnessTypeLabel = forms.label(form, "Fitness Type:", xvalue, yvalue) --Name of fitness
-FitnessAmountLabel = forms.label(form, "Amount", incrementX(AmountX), yvalue) --How much each fitness is worth
-FitnessCheckLabel = forms.label(form, "On", incrementX(AmountX), yvalue) --Whether the fitness timeout constant can be reseted with this fitness.
-FitnessTimeoutLabel = forms.label(form, "Timeout", incrementX(AmountX), yvalue) --Enable this fitnyvaless
-yvalue=incrementY(AmountY) xvalue=0
---Rightmost Options
-RightmostLabel = forms.label(form, "Rightmost ", xvalue, yvalue) --Rightmost Label is the fitness for how far right you can go
-RightmostAmount = forms.textbox(form, 1, 60, 20, nil, incrementX(AmountX), yvalue) --Each pixel right move is multiplied by this number
-RightmostFitness = forms.checkbox(form, "", incrementX(AmountX), yvalue, "true") --If an organism reaches farther than right than ever before during a generation reset the timeout constant
-RightmostTimeout = forms.checkbox(form, "", incrementX(AmountX), yvalue, "true")--Toggle the rightmost fitness type
-yvalue=incrementY(AmountY) xvalue=0
---Novelty Options
-NoveltyLabel = forms.label(form, "Novelty ", xvalue, yvalue) --Novelty Label is the fitness for how unique an orgranism travels
-NoveltyAmount = forms.textbox(form, 10000, 60, 20, nil, incrementX(AmountX), yvalue)--Each spot an orgranism goes to that No more than the Novelty Constant goes to gets this many points
-NoveltyFitness = forms.checkbox(form, "", incrementX(AmountX), yvalue) --Each new place an orgranism goes resets the timeout
-NoveltyTimeout = forms.checkbox(form, "", incrementX(AmountX), yvalue, "true") --Toggle the Novelty fitness type
-yvalue=incrementY(AmountY) xvalue=0
---Score Amount
-ScoreLabel = forms.label(form, "Score ", xvalue, yvalue)--Score Label is the fitness for how much score an organism gains during a run
-ScoreAmount = forms.textbox(form, 1, 60, 20, nil, incrementX(AmountX), yvalue)--The Score is multiplied by this number
-ScoreFitness = forms.checkbox(form, "", incrementX(AmountX), yvalue, "true") --Each time the score changes an orgranism resets there constant
-yvalue=incrementY(AmountY) xvalue=0
--- Round Amount
-RoundLabel = forms.label(form, "Round Amount ", xvalue, yvalue)
-RoundAmountValue = forms.textbox(form, RoundAmountConstant, 60, 20, nil, incrementX(AmountX), yvalue)
-RoundAmountFitness = forms.checkbox(form, "", incrementX(AmountX), yvalue,"true")
-yvalue=incrementY(AmountY) xvalue=0
--- Novelty and Timeout Constants
-NoveltyLabel = forms.label(form, "Novelty Const: ", xvalue, yvalue)
-NoveltyConstantText = forms.textbox(form, NoveltyConstant, 30, 20, nil, incrementX(AmountX), yvalue) --How many orgranism can visit a spot and it still be unique
-TimeoutLabel = forms.label(form, "Timeout Const: ", incrementX(AmountX), yvalue) --How many frames till an orgranism dies off if not reset by a fitness
-TimeoutConstantText = forms.textbox(form, TimeoutConstant, 30, 20, nil, incrementX(AmountX), yvalue) 
-yvalue=incrementY(AmountY) xvalue=0
---Continous Play and Death
-showContinousPlay = forms.checkbox(form, "Continous Play", xvalue, yvalue,"true")--Play from where the last orgranism left off
-showContinousDeath = forms.checkbox(form, "Continous Death", incrementX(AmountX), yvalue)--Play from where the last orgranism left off
-showTraining = forms.checkbox(form, "Training Coming Soon", incrementX(AmountX), yvalue)--Play from where the last orgranism left off
-yvalue=incrementY(AmountY) xvalue=0
-saveButton = forms.button(form, "Save", savePool, xvalue, yvalue) --Save the Network
-loadButton = forms.button(form, "Load", loadPool, incrementX(AmountX), yvalue) --Load the Network
-restartButton = forms.button(form, "Restart", initializePool, incrementX(AmountX), yvalue) --Restart the experiment
-loadCurrentButton = forms.button(form, "Load Current", loadCurrent, incrementX(AmountX), yvalue) --Load the Network
-yvalue=incrementY(AmountY) xvalue=0
---File Save
-saveLoadLabel = forms.label(form, "Save/Load:", xvalue, yvalue)
-saveLoadFile = forms.textbox(form, Filename .. ".pool", 110, 25, nil, incrementX(AmountX), yvalue)
---playTopButton = forms.button(form, "Play Top", playTop, incrementX(AmountX), yvalue)
---Hides banner
--- hideBanner = forms.checkbox(form, "Hide Banner", incrementX(AmountX), yvalue)
--- yvalue=incrementY(AmountY)
--- xvalue=0
---What you are going to name the file
-
-end
 
 FitnessBox(140,40,600,700)
-
 
 
 writeFile("temp.pool")
